+++ ErsatzTV/Services/SchedulerService.cs
@@ -213,18 +213,13 @@ public class SchedulerService : BackgroundService
             .Include(p => p.Channel)
             .ToListAsync(cancellationToken);
 
-        foreach (Playout playout in playouts.OrderBy(p => decimal.Parse(p.Channel.Number, CultureInfo.InvariantCulture)))
+        foreach (Playout playout in playouts.OrderBy(p => decimal.Parse(
+                     p.Channel.Number,
+                     CultureInfo.InvariantCulture)))
         {
             await _workerChannel.WriteAsync(
                 new BuildPlayout(playout.Id, PlayoutBuildMode.Continue),
                 cancellationToken);
-
-            if (playout.Channel.PlayoutMode is ChannelPlayoutMode.OnDemand)
-            {
-                await _workerChannel.WriteAsync(
-                    new TimeShiftOnDemandPlayout(playout.Id, DateTimeOffset.Now, false),
-                    cancellationToken);
-            }
         }
     }